third_party/WebKit/LayoutTests/fast/forms/file/file-input-webkitdirectory-key-space-expected.txt

CONSOLE MESSAGE: line 5: FileChooser: opened; mode=kUploadFolder
CONSOLE MESSAGE: line 5: FileChooser: mock-file-chooser.js doesn't support directory selection yet.
Ensure pressing Space key with focus on an INPUT element with TYPE=FILE and WEBKITDIRECTORY launches a file chooser. Automated test passes if 'FileChooser: opened' was logged.
